pyproject.toml
[tool.poetry]
name = "service_auth"
version = "0.1.0"
description = ""
authors = ["Indoc Research <admin@indocresearch.org>"]

[tool.poetry.dependencies]
python = "^3.7"
aniso8601 = "8.0.0"
casbin = "1.1.1"
casbin_sqlalchemy_adapter = "0.3.1"
certifi = "^2021.10.8"
cffi = "1.15.0"
chardet = "3.0.4"
click = "7.1.2"
common = "0.0.20"
docopt = "0.6.2"
gunicorn = "20.0.4"
hdfs = "2.5.8"
idna = "2.9"
itsdangerous = "1.1.0"
Jinja2 = "2.11.2"
MarkupSafe = "1.1.1"
neo4j = "4.0.0"
opentelemetry-exporter-jaeger = "1.6.2"
opentelemetry-instrumentation = "0.26b1"
opentelemetry-instrumentation-flask = "0.26b1"
opentelemetry-instrumentation-logging = "0.26b1"
opentelemetry-instrumentation-requests = "0.26b1"
opentelemetry-instrumentation-sqlalchemy = "0.26b1"
psycopg2-binary = "2.9.3"
pycparser = "2.20"
pydantic = "1.8.2"
PyJWT = "1.4.2"
python-json-logger = "0.1.11"
python-keycloak = "0.21.0"
python-ldap = "3.3.1"
pytz = "2020.1"
requests = "2.24.0"
six = "1.15.0"
SQLAlchemy = "1.3.17"
urllib3 = "1.25.9"
Werkzeug = "1.0.1"
fastapi = "^0.73.0"
opentelemetry-instrumentation-fastapi = "0.26b1"
fastapi-utils = "^0.2.1"
uvicorn = "^0.17.4"
packaging = "^21.3"
importlib-metadata = "1.7.0"
httpx = "^0.22.0"
fastapi-sqlalchemy = "0.2.1"
logger = "0.1.0"

[tool.poetry.dev-dependencies]
pytest-httpx = "^0.20.0"
pytest = "^7.0.1"
pytest-cov = "^3.0.0"
pytest-env = "^0.6.2"
pytest-mock = "^3.7.0"
mock = "4.0.3"
sqlalchemy-utils = "0.38.2"
testcontainers = "3.4.2"


[tool.pytest.ini_options]
testpaths = "tests"
addopts = "--cov logger --cov-report term"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "pilot"
url = "https://git.indocresearch.org/api/v4/groups/pilot/-/packages/pypi/simple"
default = true
